Remove common.ts from handlers, move functions to utils.ts
