+++ src/entrypoints/content/modules/skill-bonus/index.ts
@@ -9,46 +9,8 @@ import {
 } from '@/entrypoints/content/common/utils/dom'
 import { isCurrentPage, pages } from '@/entrypoints/content/common/utils/pages'
 import metadata from '@/entrypoints/content/modules/skill-bonus/metadata'
-import { calcBonuses, SkillBonuses } from '@/entrypoints/content/modules/skill-bonus/utils'
-import { i18n } from '#i18n'
+import { calcBonuses, createBonusBars } from '@/entrypoints/content/modules/skill-bonus/utils'
 
-const MAX_BAR_LENGTH = 20
-
-/**
- * Create a skill bar element for the bonus.
- *
- * @param skillBar - The skill bar element
- * @param bonuses - The bonuses to display
- * @returns An array of div elements representing bonus skill bars
- */
-const createBonusBars = (skillBar: HTMLDivElement, bonuses: SkillBonuses): HTMLDivElement[] => {
-  const level = parseInt(skillBar.getAttribute('level') ?? '0', 10)
-  const denominationBar = querySelectorIn<HTMLSpanElement>(skillBar, '.bar-max > .bar-denomination')
-  if (!level || !denominationBar) return []
-
-  const bonusBars: HTMLDivElement[] = []
-  let accumulatedBonus = 0
-
-  ;(Object.entries(bonuses) as [keyof SkillBonuses, number][]).forEach(([bonusType, bonusValue]) => {
-    if (bonusValue === 0) return
-
-    accumulatedBonus += bonusValue
-
-    const bonusBar = document.createElement('div')
-    bonusBar.className = `hte-skill-bonus-bar hte-skill-bonus-bar-${bonusType}`
-    bonusBar.style.width = `${Math.round(((level + accumulatedBonus) / MAX_BAR_LENGTH) * 100)}%`
-    bonusBar.title = i18n.t(`skill_bonus_${bonusType}`, [bonusValue.toFixed(2)])
-    bonusBar.appendChild(denominationBar.cloneNode(true))
-
-    bonusBars.push(bonusBar)
-  })
-
-  return bonusBars
-}
-
-/**
- * Extend player skill bars to include loyalty and homegrown bonuses.
- */
 const skillBonus: Module = {
   metadata,
   pages: [pages.playerDetail.senior.own, pages.playerList.senior.own],
@@ -57,7 +19,6 @@ const skillBonus: Module = {
 
     if (isCurrentPage(pages.playerDetail.senior.own)) {
       const element = querySelector('#mainBody .playerInfo')
-
       if (element) elements.push(element)
     } else if (isCurrentPage(pages.playerList.senior.own)) {
       elements.push(...Array.from(querySelectorAll('#mainBody > .playerList > .teamphoto-player')))
@@ -71,7 +32,7 @@ const skillBonus: Module = {
 
       skillBars.forEach((skillBar) => {
         const levelBar = querySelectorIn<HTMLDivElement>(skillBar, '.bar-level', false)
-        if (!levelBar) return // non-existent skills won't have a .bar-level
+        if (!levelBar) return
 
         const bonusBars = createBonusBars(skillBar, bonuses)
         let insertionPoint = levelBar
