+++ src/entrypoints/content/modules/skill-bonus/utils.ts
@@ -1,5 +1,8 @@
 import { querySelectorAllIn, querySelectorIn } from '@/entrypoints/content/common/utils/dom'
 import { pages } from '@/entrypoints/content/common/utils/pages'
+import { i18n } from '#i18n'
+
+const MAX_BAR_LENGTH = 20
 
 const MAX_LOYALTY = 20
 
@@ -19,6 +22,31 @@ export const calcBonuses = (element: Element): SkillBonuses => {
     loyalty: calcLoyaltyBonus(element),
     homegrown: calcHomegrownBonus(element),
   }
+}
+
+export const createBonusBars = (skillBar: HTMLDivElement, bonuses: SkillBonuses): HTMLDivElement[] => {
+  const level = parseInt(skillBar.getAttribute('level') ?? '0', 10)
+  const denominationBar = querySelectorIn<HTMLSpanElement>(skillBar, '.bar-max > .bar-denomination')
+  if (!level || !denominationBar) return []
+
+  const bonusBars: HTMLDivElement[] = []
+  let accumulatedBonus = 0
+
+  ;(Object.entries(bonuses) as [keyof SkillBonuses, number][]).forEach(([bonusType, bonusValue]) => {
+    if (bonusValue === 0) return
+
+    accumulatedBonus += bonusValue
+
+    const bonusBar = document.createElement('div')
+    bonusBar.className = `hte-skill-bonus-bar hte-skill-bonus-bar-${bonusType}`
+    bonusBar.style.width = `${Math.round(((level + accumulatedBonus) / MAX_BAR_LENGTH) * 100)}%`
+    bonusBar.title = i18n.t(`skill_bonus_${bonusType}`, [bonusValue.toFixed(2)])
+    bonusBar.appendChild(denominationBar.cloneNode(true))
+
+    bonusBars.push(bonusBar)
+  })
+
+  return bonusBars
 }
 
 /**
